Refactor remotes/docker errors logic
