+++ core/remotes/docker/errors/errcode.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package docker
+package errors
 
 import (
 	"encoding/json"
