+++ core/remotes/docker/errors/errdesc.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package docker
+package errors
 
 import (
 	"fmt"
